Design changes.
Changed the “menu” in the navbar to the meta color.
Re-added Plex Mono Bold font. Using it on the article headings.
Removed the grater than icon in the footer.
Replaced the cursor from laying to standing.
Updated the terminal icon to use the new icons across the board.
Added the avatar for future use.